requirements-dev.txt
# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Testing
pytest==9.0.1
pytest-django==4.11.1
pytest-cov==7.0.0
coverage==7.11.3  
factory-boy==3.3.3  

# Code quality
flake8==7.3.0
black==25.11.0
isort==7.0.0
pylint==4.0.2

# Security
bandit==1.8.6
safety==3.7.0

# Debug
django-debug-toolbar==6.1.0  
